requirements.txt
Flask
flask-cors
pymongo
python-dotenv
bcrypt
requests
langdetect
sentence-transformers
gunicorn